+++ configuration/src/configs/database.rs
@@ -25,6 +25,7 @@ pub struct DatabaseConfig {
     // Migrations cannot be applied to read-only replicas
     // We should run rpc-server only on read-only replicas
     pub read_only: bool,
+    pub shard_layout: Option<near_primitives::shard_layout::ShardLayout>,
 }
 
 impl DatabaseConfig {
@@ -34,6 +35,7 @@ impl DatabaseConfig {
             shards_config: self.shards_config.clone(),
             max_connections: self.max_connections,
             read_only: true,
+            shard_layout: self.shard_layout.clone(),
         }
     }
 }
@@ -76,6 +78,7 @@ impl From<CommonDatabaseConfig> for DatabaseConfig {
                 .max_connections
                 .unwrap_or_else(CommonDatabaseConfig::default_max_connections),
             read_only: false,
+            shard_layout: crate::shard_layout().ok(),
         }
     }
 }