+++ state-indexer/src/main.rs
@@ -23,15 +23,15 @@ async fn main() -> anyhow::Result<()> {
     if let Some(auth_token) = &indexer_config.general.rpc_auth_token {
         rpc_client = rpc_client.header(near_jsonrpc_client::auth::Authorization::bearer(auth_token)?);
     }
-    let genesis_config = rpc_client
-        .call(near_jsonrpc_client::methods::EXPERIMENTAL_genesis_config::RpcGenesisConfigRequest)
-        .await?;
-    let shard_layout = configs::shard_layout(genesis_config).await?;
+
+    let shard_layout = indexer_config
+        .database
+        .shard_layout
+        .clone()
+        .expect("Shard Layout is missing in the config.");
     let near_client = logic_state_indexer::NearJsonRpc::new(rpc_client);
 
-    let db_manager =
-        database::prepare_db_manager::<database::PostgresDBManager>(&indexer_config.database, shard_layout.clone())
-            .await?;
+    let db_manager = database::prepare_db_manager::<database::PostgresDBManager>(&indexer_config.database).await?;
     let start_block_height = configs::get_start_block_height(
         &near_client,
         &db_manager,